requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.3.5
langchain==0.0.335
chromadb==0.4.18
python-multipart==0.0.6
pydub==0.25.1
numpy==1.26.2
pytest==7.4.3
httpx==0.25.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
